refactor: 층 관련 조회 API shared 승격
