+++ src/pages/floorPlans/api/mapGraphApi.ts
@@ -1,67 +1,17 @@
 import type {
   CreateMapEdgeRequest,
   CreateMapNodeRequest,
-  FloorGraphResponse,
   MapEdgeResponse,
   MapNodeResponse,
   UpdateMapNodePositionRequest,
 } from '@apis/__generated__/data-contracts';
 import { request as apiRequest, HTTP_METHOD } from '@apis/config/request';
 import { API_ENDPOINTS } from '@apis/constants/endpoints';
+import { toMapEdge, toMapNode } from '@apis/floors/mapGraphApi';
+import type { MapEdge, MapNode } from '@apis/floors/mapGraphApi';
 
-export type MapNodeType = 'STAIR' | 'ROOM' | 'HALLWAY' | 'DOOR' | 'EXIT' | 'CUSTOM';
-
-export interface MapNode {
-  id: string;
-  code: string;
-  type: MapNodeType;
-  name: string;
-  /** 0~1 비율 좌표 */
-  x: number;
-  /** 0~1 비율 좌표 */
-  y: number;
-  isExitTarget: boolean;
-}
-
-export interface MapEdge {
-  id: string;
-  fromNodeId: string;
-  toNodeId: string;
-  distance: number;
-  bidirectional: boolean;
-}
-
-export interface FloorGraph {
-  nodes: MapNode[];
-  edges: MapEdge[];
-}
-
-const toMapNode = (response: MapNodeResponse): MapNode => {
-  const { id, code, type, name, x, y } = response;
-  if (!id || !code || !type || !name || x === undefined || y === undefined) {
-    throw new Error('노드 응답에 필수 필드가 누락되었습니다.');
-  }
-  return { id, code, type, name, x, y, isExitTarget: response.isExitTarget ?? false };
-};
-
-const toMapEdge = (response: MapEdgeResponse): MapEdge => {
-  const { id, fromNodeId, toNodeId, distance, bidirectional } = response;
-  if (!id || !fromNodeId || !toNodeId || distance === undefined || bidirectional === undefined) {
-    throw new Error('엣지 응답에 필수 필드가 누락되었습니다.');
-  }
-  return { id, fromNodeId, toNodeId, distance, bidirectional };
-};
-
-export async function getFloorGraph(floorId: string): Promise<FloorGraph> {
-  const graph = await apiRequest<FloorGraphResponse>({
-    method: HTTP_METHOD.GET,
-    url: API_ENDPOINTS.MAP_GRAPH.DETAIL(floorId),
-  });
-  return {
-    nodes: (graph.nodes ?? []).map(toMapNode),
-    edges: (graph.edges ?? []).map(toMapEdge),
-  };
-}
+export { getFloorGraph } from '@apis/floors/mapGraphApi';
+export type { FloorGraph, MapEdge, MapNode, MapNodeType } from '@apis/floors/mapGraphApi';
 
 export async function createMapNode(floorId: string, body: CreateMapNodeRequest): Promise<MapNode> {
   const node = await apiRequest<MapNodeResponse, CreateMapNodeRequest>({
